d Baton Rouge, LA, 70803 September 24th, 2020
Lucid
New Orleans, LA

Dear Recruiting Manager,
I am writing to apply for the Summer 2021 tech internship. Currently I
am pursuing a bachelorâs degree in computer science, with an
expected graduation date of May 2022. I am very interested in working
with big data, and I believe my experience developing software in a team
environment would make me a perfect fit for this position.
As a computer science major at LSU I was able to develop an Android app
for my Object-Oriented Design class. I worked among a team of 6 students
to create a music-discovery app called âSparkâ using Java and
Android Studio. The idea behind the app was to make music discovery fun,
easy, and social. We implemented an interface similar to Tinder, where
the app would display a song and you could swipe right or left depending
on whether you liked or disliked it. To give the app a social aspect, we
also added a swipe up option that allowed you to send a song to a
friend. We did all this remotely, hosting meetings via zoom to design
the app and update each other on our work. Through this experience, I
gained a better understanding of the advantages of collaboration, and
how to manage a project with multiple people.
This past summer I developed an Android game using Unity. For this
project, I had to design all the UI and software architecture myself. It
was a very rewarding experience, seeing my idea go from a set of paper
design diagrams to a fully-realized app published on the Google Play
Store. By doing everything myself, I learned a lot about what truly goes
into the lifecycle of an application, and how to manage the ups and
downs of development.
I am very interested in working at Lucid, and I know my experience
working with a team to developing software would provide a valuable
asset to your company. If you have any questions feel free to contact me
at ichen2@lsu.edu. Thank you for your consideration, and I hope to heari